+++ src/parse.h
@@ -41,7 +41,7 @@ typedef char LINE_BUFFER[MAX_LINE_LENGTH+1];
 
 struct ASSIGNMENT
 {
-	int base;
+	long long int base;
 	unsigned int exponent;
 	int bit_min;
 	int bit_max;
@@ -53,14 +53,14 @@ struct ASSIGNMENT
 
 struct SELFTEST
 {
-	int base;
+	long long int base;
 	unsigned int exponent;
 	int bit_min;
 	unsigned long long k;
 };
 
-int valid_assignment(int base, unsigned int exp, int bit_min, int bit_max, int verbosity);	// nonzero if assignment is valid
-enum ASSIGNMENT_ERRORS get_next_assignment(char *filename, int *base, unsigned int *exponent, int *bit_min, int *bit_max, LINE_BUFFER *assignment_key, int verbosity);
-enum ASSIGNMENT_ERRORS clear_assignment(char *filename, int base, unsigned int exponent, int bit_min, int bit_max, int bit_min_new);
+int valid_assignment(long long int base, unsigned int exp, int bit_min, int bit_max, int verbosity);	// nonzero if assignment is valid
+enum ASSIGNMENT_ERRORS get_next_assignment(char *filename, long long int *base, unsigned int *exponent, int *bit_min, int *bit_max, LINE_BUFFER *assignment_key, int verbosity);
+enum ASSIGNMENT_ERRORS clear_assignment(char *filename, long long int base, unsigned int exponent, int bit_min, int bit_max, int bit_min_new);
 int process_add_file(char *workfile, char *addfile, int *addfilesstatus, int verbosity);
-enum ASSIGNMENT_ERRORS get_next_selftest(FILE* f_in, char *filename, int *base, unsigned int *exponent, int *bit_min, unsigned long long *k, int verbosity);
+enum ASSIGNMENT_ERRORS get_next_selftest(FILE* f_in, char *filename, long long int *base, unsigned int *exponent, int *bit_min, unsigned long long *k, int verbosity);
